Portability update to the ref10 library.
- The Makefiles now create both static and dynamic libraries.
- CFFI modules now statically link, removing the need to set LD_LIBRARY_PATH.
- Increased portability by using stdint.h for fixed-size integers.
- Increased portability by adding kernelrandombytes implementations for Windows and Mac OS.

+++ ref10/include/crypto_int64.h
@@ -1,6 +1,8 @@
 #ifndef crypto_int64_h
 #define crypto_int64_h
 
-typedef long long crypto_int64;
+#include <stdint.h>
+
+typedef int64_t crypto_int64;
 
 #endif
